requirements.txt
PyOpenGL
PyOpenGL_accelerate
PySDL2
numpy
Pillow